+++ pkg/sql/sem/tree/casts_test.go
@@ -208,77 +208,3 @@ func TestTupleCastVolatility(t *testing.T) {
 		}
 	}
 }
-
-func TestCastStringToRegClassTableName(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	testCases := []struct {
-		in       string
-		expected TableName
-	}{
-		{"a", MakeUnqualifiedTableName("a")},
-		{`a"`, MakeUnqualifiedTableName(`a"`)},
-		{`"a""".bB."cD" `, MakeTableNameWithSchema(`a"`, "bb", "cD")},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.in, func(t *testing.T) {
-			out, err := castStringToRegClassTableName(tc.in)
-			require.NoError(t, err)
-			require.Equal(t, tc.expected, out)
-		})
-	}
-
-	errorTestCases := []struct {
-		in            string
-		expectedError string
-	}{
-		{"a.b.c.d", "too many components: a.b.c.d"},
-		{"", `invalid table name: `},
-	}
-
-	for _, tc := range errorTestCases {
-		t.Run(tc.in, func(t *testing.T) {
-			_, err := castStringToRegClassTableName(tc.in)
-			require.EqualError(t, err, tc.expectedError)
-		})
-	}
-
-}
-
-func TestSplitIdentifierList(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	testCases := []struct {
-		in       string
-		expected []string
-	}{
-		{`abc`, []string{"abc"}},
-		{`abc.dEf  `, []string{"abc", "def"}},
-		{` "aBc"  . d  ."HeLLo"""`, []string{"aBc", "d", `HeLLo"`}},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.in, func(t *testing.T) {
-			out, err := splitIdentifierList(tc.in)
-			require.NoError(t, err)
-			require.Equal(t, tc.expected, out)
-		})
-	}
-
-	errorTestCases := []struct {
-		in            string
-		expectedError string
-	}{
-		{`"unclosed`, `invalid name: unclosed ": "unclosed`},
-		{`"unclosed""`, `invalid name: unclosed ": "unclosed""`},
-		{`hello !`, `invalid name: expected separator .: hello !`},
-	}
-
-	for _, tc := range errorTestCases {
-		t.Run(tc.in, func(t *testing.T) {
-			_, err := splitIdentifierList(tc.in)
-			require.EqualError(t, err, tc.expectedError)
-		})
-	}
-}
