+++ pkg/sql/sem/tree/datum_invariants_test.go
@@ -12,15 +12,10 @@ package tree
 
 import (
 	"testing"
-	"time"
 
-	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
-	"github.com/cockroachdb/cockroach/pkg/util/timeofday"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestAllTypesCastableToString(t *testing.T) {
@@ -40,121 +35,5 @@ func TestAllTypesCastableFromString(t *testing.T) {
 		if ok, _, _ := isCastDeepValid(types.String, typ); !ok {
 			t.Errorf("%s is not castable from STRING, all types should be", typ)
 		}
-	}
-}
-
-func TestCompareTimestamps(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	defer log.Scope(t).Close(t)
-
-	pacificTimeZone := int32(7 * 60 * 60)
-	sydneyTimeZone := int32(-10 * 60 * 60)
-
-	sydneyFixedZone := time.FixedZone("otan@sydney", -int(sydneyTimeZone))
-	// kiwiFixedZone is 2 hours ahead of Sydney.
-	kiwiFixedZone := time.FixedZone("otan@auckland", -int(sydneyTimeZone)+2*60*60)
-
-	ddate, err := NewDDateFromTime(time.Date(2019, time.November, 22, 0, 0, 0, 0, time.UTC))
-	require.NoError(t, err)
-
-	testCases := []struct {
-		desc     string
-		left     Datum
-		right    Datum
-		location *time.Location
-		expected int
-	}{
-		{
-			desc:     "same DTime are equal",
-			left:     MakeDTime(timeofday.New(12, 0, 0, 0)),
-			right:    MakeDTime(timeofday.New(12, 0, 0, 0)),
-			expected: 0,
-		},
-		{
-			desc:     "same DTimeTZ are equal",
-			left:     NewDTimeTZFromOffset(timeofday.New(22, 0, 0, 0), sydneyTimeZone),
-			right:    NewDTimeTZFromOffset(timeofday.New(22, 0, 0, 0), sydneyTimeZone),
-			expected: 0,
-		},
-		{
-			desc:     "DTime and DTimeTZ both UTC, and so are equal",
-			left:     MakeDTime(timeofday.New(12, 0, 0, 0)),
-			right:    NewDTimeTZFromOffset(timeofday.New(12, 0, 0, 0), 0),
-			expected: 0,
-		},
-		{
-			desc:     "DTime and DTimeTZ both Sydney time, and so are equal",
-			left:     MakeDTime(timeofday.New(12, 0, 0, 0)),
-			right:    NewDTimeTZFromOffset(timeofday.New(12, 0, 0, 0), sydneyTimeZone),
-			location: sydneyFixedZone,
-			expected: 0,
-		},
-		{
-			desc:     "DTimestamp and DTimestampTZ (Sydney) equal in Sydney zone",
-			left:     MustMakeDTimestamp(time.Date(2019, time.November, 22, 10, 0, 0, 0, time.UTC), time.Microsecond),
-			right:    MustMakeDTimestampTZ(time.Date(2019, time.November, 22, 10, 0, 0, 0, sydneyFixedZone), time.Microsecond),
-			location: sydneyFixedZone,
-			expected: 0,
-		},
-		{
-			desc:     "DTimestamp and DTimestampTZ (Sydney) equal in Sydney+2 zone",
-			left:     MustMakeDTimestamp(time.Date(2019, time.November, 22, 12, 0, 0, 0, time.UTC), time.Microsecond),
-			right:    MustMakeDTimestampTZ(time.Date(2019, time.November, 22, 10, 0, 0, 0, sydneyFixedZone), time.Microsecond),
-			location: kiwiFixedZone,
-			expected: 0,
-		},
-		{
-			desc:     "Date and DTimestampTZ (Sydney) equal in Sydney zone",
-			left:     ddate,
-			right:    MustMakeDTimestampTZ(time.Date(2019, time.November, 22, 0, 0, 0, 0, sydneyFixedZone), time.Microsecond),
-			location: sydneyFixedZone,
-			expected: 0,
-		},
-		{
-			desc:     "Date and DTimestampTZ (Sydney) equal in Sydney+2 zone",
-			left:     ddate,
-			right:    MustMakeDTimestampTZ(time.Date(2019, time.November, 21, 22, 0, 0, 0, sydneyFixedZone), time.Microsecond),
-			location: kiwiFixedZone,
-			expected: 0,
-		},
-		{
-			desc:     "equal wall clock time for DTime and DTimeTZ, with TimeTZ ahead",
-			left:     MakeDTime(timeofday.New(12, 0, 0, 0)),
-			right:    NewDTimeTZFromOffset(timeofday.New(22, 0, 0, 0), sydneyTimeZone),
-			expected: 1,
-		},
-		{
-			desc:     "equal wall clock time for DTime and DTimeTZ, with TimeTZ behind",
-			left:     MakeDTime(timeofday.New(12, 0, 0, 0)),
-			right:    NewDTimeTZFromOffset(timeofday.New(5, 0, 0, 0), pacificTimeZone),
-			expected: -1,
-		},
-		{
-			desc:     "equal wall clock time for DTime and DTimeTZ, with TimeTZ ahead",
-			left:     NewDTimeTZFromOffset(timeofday.New(22, 0, 0, 0), sydneyTimeZone),
-			right:    NewDTimeTZFromOffset(timeofday.New(5, 0, 0, 0), pacificTimeZone),
-			expected: -1,
-		},
-		{
-			desc:     "wall clock time different for DTimeTZ and DTimeTZ",
-			left:     NewDTimeTZFromOffset(timeofday.New(23, 0, 0, 0), sydneyTimeZone),
-			right:    NewDTimeTZFromOffset(timeofday.New(5, 0, 0, 0), pacificTimeZone),
-			expected: 1,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(
-			tc.desc,
-			func(t *testing.T) {
-				ctx := &EvalContext{
-					SessionData: &sessiondata.SessionData{
-						Location: tc.location,
-					},
-				}
-				assert.Equal(t, tc.expected, compareTimestamps(ctx, tc.left, tc.right))
-				assert.Equal(t, -tc.expected, compareTimestamps(ctx, tc.right, tc.left))
-			},
-		)
 	}
 }
